Split span token read/init logic
