+++ test/test_contrib/test_github_wiki.py
@@ -1,29 +1,21 @@
-from unittest import TestCase, mock
+from mistletoe.base_elements import serialize_tokens
 from mistletoe.span_tokenizer import tokenize_span
-from mistletoe.parse_context import get_parse_context
-from contrib.github_wiki import GithubWiki, GithubWikiRenderer
+from contrib.github_wiki import GithubWikiRenderer
 
 
-class TestGithubWiki(TestCase):
-    def setUp(self):
-        self.renderer = GithubWikiRenderer()
-        self.renderer.__enter__()
-        self.addCleanup(self.renderer.__exit__, None, None, None)
+def test_parse(data_regression):
+    with GithubWikiRenderer():
+        source = "text with [[wiki | target]]"
+        data_regression.check(serialize_tokens(tokenize_span(source), as_dict=True))
 
-    def test_parse(self):
-        MockRawText = mock.Mock(autospec="mistletoe.span_tokens.RawText")
-        RawText = get_parse_context().span_tokens.pop()
-        get_parse_context().span_tokens.append(MockRawText)
-        try:
-            tokens = tokenize_span("text with [[wiki | target]]")
-            token = tokens[1]
-            self.assertIsInstance(token, GithubWiki)
-            self.assertEqual(token.target, "target")
-            MockRawText.assert_has_calls([mock.call("text with "), mock.call("wiki")])
-        finally:
-            get_parse_context().span_tokens[-1] = RawText
 
-    def test_render(self):
-        token = next(iter(tokenize_span("[[wiki|target]]")))
-        output = '<a href="target">wiki</a>'
-        self.assertEqual(self.renderer.render(token), output)
+def test_parse_with_children(data_regression):
+    with GithubWikiRenderer():
+        source = "[[*alt*|link]]"
+        data_regression.check(serialize_tokens(tokenize_span(source), as_dict=True))
+
+
+def test_render(file_regression):
+    with GithubWikiRenderer() as renderer:
+        token = tokenize_span("[[wiki|target]]")[0]
+        file_regression.check(renderer.render(token), extension=".html")
