+++ test/test_contrib/test_github_wiki/test_parse.yml
@@ -1,0 +1,7 @@
+- RawText:
+    content: 'text with '
+- GithubWiki:
+    children:
+    - RawText:
+        content: wiki
+    target: target
