+++ test/test_contrib/test_github_wiki/test_parse_with_children.yml
@@ -1,0 +1,7 @@
+- GithubWiki:
+    children:
+    - Emphasis:
+        children:
+        - RawText:
+            content: alt
+    target: link
