+++ test/test_contrib/test_github_wiki/test_render.html
@@ -1,0 +1,1 @@
+<a href="target">wiki</a>
